Add config flag. Increase default value.
