+++ x-pack/plugins/ingest_manager/server/routes/index.ts
@@ -14,4 +14,4 @@ export { registerRoutes as registerInstallScriptRoutes } from './install_script'
 export { registerRoutes as registerOutputRoutes } from './output';
 export { registerRoutes as registerSettingsRoutes } from './settings';
 export { registerRoutes as registerAppRoutes } from './app';
-export { preAuthHandler } from './global_interceptors';
+export { registerLimitedConcurrencyRoutes } from './limited_concurrency';
